# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UiRealtyPals API",
    "description": "Full REST API for UiRealtyPals - AI-powered real estate discovery platform for Noida. Auth: users send Authorization: Bearer <supabase-access-token>. Admin uses admin_session cookie. Webhooks use X-Webhook-Secret header.",
    "version": "1.0.0",
    "contact": {
      "name": "RealtyPals Engineering",
      "url": "https://realty-pals-dev-frontend-three.vercel.app"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local dev server"
    },
    {
      "url": "https://realtypalsdev.onrender.com",
      "description": "Production (Render)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase access token. Obtain via supabase.auth.signIn*(). Send as: Authorization: Bearer <token>"
      },
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session",
        "description": "Admin session cookie set by POST /api/v1/admin/login. Required for all /admin/* routes."
      },
      "WebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Webhook-Secret",
        "description": "HMAC-SHA256 secret for webhook lead ingestion."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Not found"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "storage_url": {
            "type": "string",
            "format": "uri"
          },
          "doc_type": {
            "type": "string",
            "enum": [
              "rera",
              "brochure",
              "floor_plan",
              "price_list",
              "payment_plan",
              "legal",
              "other"
            ]
          },
          "file_size_bytes": {
            "type": "integer",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "parameters": {
      "SlugPath": {
        "name": "slug",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "URL-safe slug identifier"
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized - valid Bearer token or admin session required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request body or query parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Too many requests - IP rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/projects/{slug}/documents": {
      "get": {
        "tags": [
          "Projects",
          "Documents"
        ],
        "summary": "List project documents",
        "description": "RERA, brochures, floor plans. Ordered newest first.",
        "operationId": "getProjectDocuments",
        "parameters": [
          {
            "$ref": "#/components/parameters/SlugPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "summary": "List documents for project",
        "description": "Public. Pass ?slug=<project-slug>.",
        "operationId": "listDocuments",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Docs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "docs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Document"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upload project document (admin)",
        "description": "Admin only. Validates file by magic bytes (not just MIME). For images: extracts text via Groq vision. Orphaned files cleaned on DB error. Max 20MB. Allowed: PDF, JPEG, PNG, WebP.",
        "operationId": "uploadDocument",
        "security": [
          {
            "AdminSession": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "project_id",
                  "project_slug"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "project_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "project_slug": {
                    "type": "string"
                  },
                  "doc_type": {
                    "type": "string",
                    "enum": [
                      "rera",
                      "brochure",
                      "floor_plan",
                      "price_list",
                      "payment_plan",
                      "legal",
                      "other"
                    ],
                    "default": "other"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/documents/ask": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "AI Q&A over document",
        "description": "Grounded Q&A - will not hallucinate. Rate limited 20/min per IP.",
        "operationId": "askDocument",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "document_id",
                  "question"
                ],
                "properties": {
                  "document_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "question": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 500,
                    "example": "What is the RERA registration number?"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI answer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": {
                      "type": "string"
                    },
                    "document_name": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "502": {
            "description": "AI service unavailable"
          }
        }
      }
    },
    "/api/v1/analytics/engagement": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Track engagement event",
        "description": "Tracks drop_off, first_engagement events for conversion funnel analytics.",
        "operationId": "trackEngagement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "session_id",
                  "event"
                ],
                "properties": {
                  "session_id": {
                    "type": "string"
                  },
                  "event": {
                    "type": "string",
                    "example": "drop_off"
                  },
                  "project_id": {
                    "type": "string"
                  },
                  "drop_off_stage": {
                    "type": "string"
                  },
                  "idle_seconds": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recorded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/analytics/promotions": {
      "post": {
        "tags": [
          "Analytics"
        ],
        "summary": "Track promotion interaction",
        "description": "Tracks impression or click events for promotional banners.",
        "operationId": "trackPromotion",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action",
                  "promotional_id",
                  "session_id"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "impression",
                      "click"
                    ]
                  },
                  "promotional_id": {
                    "type": "string"
                  },
                  "session_id": {
                    "type": "string"
                  },
                  "user_id": {
                    "type": "string"
                  },
                  "guest_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tracked",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
    "/api/v1/admin/documents/{docId}": {
      "delete": {
        "summary": "Auto-generated endpoint for DELETE /api/v1/admin/documents/:docId",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "docId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/admin/projects/{id}/documents": {
      "get": {
        "summary": "Auto-generated endpoint for GET /api/v1/admin/projects/:id/documents",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        },
        "security": [
          {
            "AdminSession": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}